Migrated LED matrix and key matrix to info.json - UNTESTED
