+++ keyboards/nifty_numpad/keymaps/default/keymap.json.ignore
@@ -1,0 +1,21 @@
+{
+    "keyboard": "nifty_numpad",
+    "keymap": "default",
+    "layers": [
+        ["KC_F13", "KC_F19",    "LT(LAYER_RGB, KC_BSPC)", "RGUI(KC_TAB)", "RGUI(KC_E)", "RGUI(RSFT(KC_S))", 
+         "KC_F14", "KC_F20",    "KC_NUM",                 "KC_PSLS",      "KC_PAST",    "KC_PMNS", 
+         "KC_F15", "KC_F21",    "KC_P7",                  "KC_P8",        "KC_P9",      "KC_PPLS", 
+         "KC_F16", "KC_F22",    "KC_P4",                  "KC_P5",        "KC_P6",      
+         "KC_F17", "KC_F23",    "KC_P1",                  "KC_P2",        "KC_P3",      "KC_PENT", 
+         "KC_F18", "KC_F24",    "KC_P0",                                  "KC_PDOT"
+        ],
+        ["KC_TRNS", "KC_TRNS",    "KC_TRNS", "RGB_HUI", "RGB_SAI", "RGB_VAI", 
+         "KC_TRNS", "KC_TRNS",    "KC_TRNS", "RGB_HUD", "RGB_SAD", "RGB_VAD", 
+         "KC_TRNS", "KC_TRNS",    "KC_TRNS", "RGB_MOD", "KC_TRNS", "RGB_SPI", 
+         "KC_TRNS", "KC_TRNS",    "KC_TRNS", "KC_TRNS", "KC_TRNS", "KC_TRNS", 
+         "KC_TRNS", "KC_TRNS",    "KC_TRNS", "KC_TRNS", "RGB_SPD", "KC_TRNS", 
+         "KC_TRNS", "KC_TRNS",    "KC_TRNS"]
+    ],
+    "layout": "LAYOUT_numpad_6x6"
+}
+
