--- keyboards/nifty_numpad/keymaps/default/rules.mk
@@ -1,6 +1,0 @@
-TAP_DANCE_ENABLE=yes
-RGB_IDLE_ENABLE=yes
-
-ifeq ($(strip $(RGB_IDLE_ENABLE)), yes)
-    OPT_DEFS += -DRGB_IDLE_ENABLE
-endif
