+++ keyboards/nifty_numpad/keymaps/idle_rgb_example/rules.mk
@@ -1,0 +1,6 @@
+TAP_DANCE_ENABLE=yes
+RGB_IDLE_ENABLE=yes
+
+ifeq ($(strip $(RGB_IDLE_ENABLE)), yes)
+    OPT_DEFS += -DRGB_IDLE_ENABLE
+endif
