+++ keyboards/nifty_numpad/nifty_numpad.c
@@ -1,4 +1,4 @@
-/* Copyright 2023 irex
+/* Copyright 2023 Acliad
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
