--- keyboards/nifty_numpad/rules.mk
@@ -1,3 +1,0 @@
-RGB_MATRIX_ENABLE=yes
-RGB_MATRIX_DRIVER=WS2812
-WS2812_DRIVER=vendor
